refactor: 자동차 위치 표시바(-) 만드는 함수 Cars -> View 로 이동
